+++ include/gaia/containers/impl/sarray_impl.h
@@ -281,7 +281,7 @@ namespace std {
 	template <typename T, uint32_t N>
 	struct tuple_size<gaia::containers::sarr<T, N>>: std::integral_constant<uint32_t, N> {};
 
-	template <uint32_t I, typename T, uint32_t N>
+	template <size_t I, typename T, uint32_t N>
 	struct tuple_element<I, gaia::containers::sarr<T, N>> {
 		using type = T;
 	};
